requirements.txt
appdirs==1.4.4
capstone==5.0.7
cffi==2.0.0
cmsis-pack-manager==0.6.0
colorama==0.4.6
importlib_metadata==8.7.1
importlib_resources==6.5.2
intelhex==2.3.0
intervaltree==3.2.1
lark==1.3.1
libusb-package==1.0.26.3
natsort==8.4.0
prettytable==3.17.0
psutil==7.2.2
pycparser==3.0
pyelftools==0.32
pylink-square==1.7.0
pyocd==0.42.0
pyserial==3.5
pyusb==1.3.1
PyYAML==6.0.3
six==1.17.0
sortedcontainers==2.4.0
typing_extensions==4.15.0
wcwidth==0.6.0
zipp==3.23.0
